filebrowser 2 update

+++ inc/config.php
@@ -247,10 +247,10 @@ $btsync = processExists("resilio-sync",rslsync);
 $deluged = processExists("deluged",$username);
 $delugedweb = processExists("deluge-web",$username);
 $emby = processExists("EmbyServer",emby);
+$filebrowser = processExists("filebrowser",$username);
 $flood = processExists("flood",$username);
 $headphones = processExists("headphones",$username);
 $irssi = processExists("irssi",$username);
-$lidarr = processExists("lidarr",$username);
 $lounge = processExists("lounge",lounge);
 $nzbget = processExists("nzbget",$username);
 $nzbhydra = processExists("nzbhydra",$username);
@@ -296,11 +296,11 @@ if(file_exists('/srv/panel/custom/url.override.php')){
   $csfURL = "https://" . $_SERVER['HTTP_HOST'] . ":3443";
   $dwURL = "https://" . $_SERVER['HTTP_HOST'] . "/deluge/";
   $embyURL = "https://" . $_SERVER['HTTP_HOST'] . "/emby";
+  $filebrowserURL = "https://" . $_SERVER['HTTP_HOST'] . "/filebrowser";
   $floodURL = "https://" . $_SERVER['HTTP_HOST'] . "/flood/";
   $headphonesURL = "https://" . $_SERVER['HTTP_HOST'] . "/headphones/home";
   $jackettURL = "https://" . $_SERVER['HTTP_HOST'] . "/jackett/";
   $loungeURL = "https://" . $_SERVER['HTTP_HOST'] . "/irc";
-  $lidarrURL = "http://" . $_SERVER['HTTP_HOST'] . "/lidarr";
   $medusaURL = "https://" . $_SERVER['HTTP_HOST'] . "/medusa";
   $netdataURL = "https://" . $_SERVER['HTTP_HOST'] . "/netdata/";
   $nextcloudURL = "https://" . $_SERVER['HTTP_HOST'] . "/nextcloud";
@@ -319,7 +319,7 @@ if(file_exists('/srv/panel/custom/url.override.php')){
   $sonarrURL = "https://" . $_SERVER['HTTP_HOST'] . "/sonarr";
   $subsonicURL = "https://" . $_SERVER['HTTP_HOST'] . "/subsonic";
   $syncthingURL = "https://" . $_SERVER['HTTP_HOST'] . "/syncthing/";
-  $lidarrURL = "http://" . $_SERVER['HTTP_HOST'] . "/lidarr";
+  $lidarrURL = "https://" . $_SERVER['HTTP_HOST'] . "/lidarr";
   if ($zssl == "true") { $zncURL = "https://" . $_SERVER['HTTP_HOST'] . ":$zport"; }
   if ($zssl == "false") { $zncURL = "http://" . $_SERVER['HTTP_HOST'] . ":$zport"; }
  }
@@ -353,6 +353,8 @@ case 0:
     $cbodycp .= $couchpotato;
   $emby = isEnabled("emby-server", $username);
     $cbodye .= $emby;
+  $filebrowser = isEnabled("filebrowser", $username);
+    $cbodyfileb .= $filebrowser;
   $flood = isEnabled("flood", $username);
     $cbodyf .= $flood;
   $headphones = isEnabled("headphones", $username);
@@ -407,7 +409,10 @@ break;
 /* enable & start services */
 case 66:
   $process = $_GET['serviceenable'];
-    if ($process == "resilio-sync"){
+ if ($process == "filebrowser"){
+      shell_exec("sudo systemctl enable $process");
+      shell_exec("sudo systemctl start $process");
+    } elseif ($process == "resilio-sync"){
       shell_exec("sudo systemctl enable $process");
       shell_exec("sudo systemctl start $process");
     } elseif ($process == "shellinabox"){
@@ -471,7 +476,10 @@ break;
 /* disable & stop services */
 case 77:
   $process = $_GET['servicedisable'];
-    if ($process == "resilio-sync"){
+  if ($process == "filebrowser"){
+      shell_exec("sudo systemctl stop $process");
+      shell_exec("sudo systemctl disable $process");
+    } elseif ($process == "resilio-sync"){
       shell_exec("sudo systemctl stop $process");
       shell_exec("sudo systemctl disable $process");
     } elseif ($process == "shellinabox"){
@@ -514,7 +522,10 @@ break;
 /* restart services */
 case 88:
   $process = $_GET['servicestart'];
-    if ($process == "resilio-sync"){
+  if ($process == "filebrowser"){
+      shell_exec("sudo systemctl enable $process");
+      shell_exec("sudo systemctl restart $process");
+    } elseif ($process == "resilio-sync"){
       shell_exec("sudo systemctl enable $process");
       shell_exec("sudo systemctl restart $process");
     } elseif ($process == "shellinabox"){
